Change all integer types to Long and all floating point types to Double

+++ src/main/java/com/stripe/model/Event.java
@@ -21,7 +21,7 @@ public class Event extends ApiResource implements HasId {
   Long created;
   EventData data;
   Boolean livemode;
-  Integer pendingWebhooks;
+  Long pendingWebhooks;
   EventRequest request;
   String type;
